backend/core-api/pyproject.toml
[project]
name = "core-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.94.1",
    "fastapi>=0.135.3",
    "gunicorn>=25.3.0",
    "python-dotenv>=1.2.2",
    "python-json-logger>=4.1.0",
    "uvicorn>=0.44.0",
]
